[代码完善](dev_2.2.2): 开单时获取患者治疗计划列表
列表中的项目，需要获得项目的折扣价

+++ yunya-framework/yunya-common-core/src/main/java/com/yunya/framework/common/enums/TreatPlanStatusEnum.java
@@ -11,16 +11,16 @@ import java.util.Objects;
  * @since: 1.0.0
  */
 public enum TreatPlanStatusEnum {
-    UNCONFIRM("未确认", (byte)0),
-    CONFIRMED("已确认", (byte)1),
-    EXECUTING("进行中", (byte)2),
-    COMPLETED("全部完成", (byte)3),
-    TERMINATION("提前终止", (byte)4);
+    UNCONFIRM("未确认", 0),
+    CONFIRMED("已确认", 1),
+    EXECUTING("进行中", 2),
+    COMPLETED("全部完成", 3),
+    TERMINATION("提前终止", 4);
 
     private final String value;
-    private final Byte code;
+    private final Integer code;
 
-    TreatPlanStatusEnum(String value, Byte code) {
+    TreatPlanStatusEnum(String value, Integer code) {
         this.value = value;
         this.code = code;
     }
@@ -29,7 +29,7 @@ public enum TreatPlanStatusEnum {
         return value;
     }
 
-    public Byte getCode() {
+    public Integer getCode() {
         return code;
     }
