+++ yunya-remote/src/main/java/com/yunya/feign/emr/domain/query/TreatPlanRecordQuery.java
@@ -44,5 +44,5 @@ public class TreatPlanRecordQuery extends PageQuery implements Serializable {
 
     /** 治疗状态：0-未确认，1-已确认，2-进行中，3-已完成，4-提前终止*/
     @ApiModelProperty("治疗状态：0-未确认，1-已确认，2-进行中，3-已完成，4-提前终止")
-    private Collection<Byte> status;
+    private Collection<Integer> status;
 }
